+++ src/core/optionals/ffmpeg/builtins_ffmpeg_bitstream_filters.ml
@@ -169,7 +169,7 @@ let handler_getters (type a) ~(mk_params : a mk_params)
 let register_filters () =
   List.iter
     (fun ({ Avcodec.BitstreamFilter.name; codecs; options } as filter) ->
-      let args, args_parser = Builtins_ffmpeg_filters.mk_options options in
+      let args, args_parser = Ffmpeg_filter_options.mk_options options in
       let modes = modes name codecs in
       let base, module_name =
         if List.length modes > 1 then