+++ src/core/optionals/ffmpeg/ffmpeg_utils.mli
@@ -31,7 +31,6 @@ val conf_scaling_algorithm : string Dtools.Conf.t
 val liq_main_ticks_time_base : unit -> Avutil.rational
 val liq_audio_sample_time_base : unit -> Avutil.rational
 val liq_video_sample_time_base : unit -> Avutil.rational
-val liq_frame_time_base : unit -> Avutil.rational
 val liq_frame_pixel_format : Avutil.Pixel_format.t
 val liq_frame_pixel_format_with_alpha : Avutil.Pixel_format.t
 val pixel_format_has_alpha : Avutil.Pixel_format.t -> bool
@@ -89,9 +88,3 @@ module Duration : sig
   val push : 'a t -> 'a -> (int * (int * 'a) list) option
   val flush : 'a t -> int * (int * 'a) list
 end
-
-val mk_subtitle_decoder :
-  output:(?data:'a -> buffer:Decoder.buffer -> length:int -> unit -> unit) ->
-  process:('b -> int * int * 'a) ->
-  unit ->
-  [ `Flush | `Subtitle of 'b ] Ffmpeg_decoder_common.sparse_decoder